categoies

+++ todo/todo.service.js
@@ -1,22 +1,24 @@
 
 const todoController = require('./todo.controller');
 const usersController = require('../users/users.controller');
+const { get } = require('mongoose');
 
 async function gettodoByUser(id) {
     try {
         const user = await usersController.getUserById(id);
-        const todo = await todoController.getAlltodo();
-        const usertodo = todo.filter(todo => todo.user.toString() === user._id.toString());
-        if (!usertodo) {
-            return;
+        if (!user) {
+            throw new Error('User not found');
         }
-        // console.log("usertodo", usertodo);
-        return usertodo;
+        const alltodo = await todoController.getAlltodo();
+        const usertodo = alltodo.filter(todo => todo.user.toString() === user._id.toString());
+        const shareWithMe = alltodo.filter(todo => todo.sharedWith.includes(user._id.toString()));
+        return usertodo.concat(shareWithMe);
     } catch (error) {
         console.log(error);
         return error;
     }
 }
+
 
 // gettodoByUser("65e726f80da8d3f2276aa797")
 async function getOnetodo(id) {
@@ -28,15 +30,14 @@ async function getOnetodo(id) {
         return error;
     }
 }
-
 async function create(data) {
     try {
         const user = await usersController.getUserById(data.user);
         if (!user) {
-            throw error = new Error('no user')
+            throw new Error('User not found');
         }
         if (!data.todo) {
-            throw error = new Error('no todo')
+            throw new Error('Todo is required');
         }
         const todo = await todoController.createtodo(data);
         user.todo.unshift(todo._id);
@@ -48,15 +49,26 @@ async function create(data) {
     }
 }
 
-
 async function update(id, data) {
     try {
         const user = await usersController.getUserById(data.user);
-        user.todo.filter(todo => todo.toString() !== id.toString());
-        const todo = await todoController.updatetodo(id, data);
-        user.todo.push(todo._id);
-        await user.save();
-        return todo;
+        if (!user) {
+            throw new Error('User not found');
+        }
+
+        const todoToUpdate = await todoController.gettodoById(id);
+        if (!todoToUpdate) {
+            throw new Error('Todo not found');
+        }
+
+        // Ensure the user has access to update this todo
+        if (todoToUpdate.user.toString() !== user._id.toString()) {
+            throw new Error('Unauthorized to update this todo');
+        }
+
+        // Update the todo
+        const updatedTodo = await todoController.updatetodo(id, data);
+        return updatedTodo;
     } catch (error) {
         console.log(error);
         return error;
@@ -67,16 +79,18 @@ async function deltodo(id) {
     try {
         const todoToDelete = await todoController.gettodoById(id);
         const user = await usersController.getUserById(todoToDelete.user);
-
-        console.log("user before deletion", user);
-
+        if (!user) {
+            throw new Error('User not found');
+        }
+        if (!todoToDelete) {
+            throw new Error('Todo not found');
+        }
         // Use filter to create a new array without the todo to delete
-        user.todo = user.todo.filter(todo => todo.toString() !== id.toString());
-
-        console.log("user after deletion", user);
+        user.todo.pull(id); // pull מוציא את הערך מהמערך
+        await user.save();
 
         const deletedtodo = await todoController.deletetodo(id);
-        await user.save();
+
 
         return deletedtodo;
     } catch (error) {
@@ -85,5 +99,153 @@ async function deltodo(id) {
     }
 }
 
+async function getCategories(userId) {
+    try {
+        const cat = await todoController.getCategories(userId);
+        // console.log({cat});
+        return cat;
+    } catch (error) {
+        console.log(error);
+        return error;
+    }
+}
+// getCategories("660115e6fc28d3e1ced32a6d");
 
-module.exports = { gettodoByUser, getOnetodo, create, update, deltodo };
+async function share(id, data) {
+    try {
+        const user = await usersController.getUserById(data.user);
+        if (!user) {
+            throw new Error('User not found');
+        }
+        if (!data.todo) {
+            throw new Error('Todo is required');
+        }
+        const todo = await todoController.gettodoById(id);
+        if (!todo) {
+            throw new Error('Todo not found');
+        }
+        // Ensure the user has access to update this todo
+        if (todo.user.toString() !== user._id.toString()) {
+            throw new Error('Unauthorized to update this todo');
+        }
+        // Update the todo
+        todo.sharedWith.push(data.user);
+        await todo.save();
+        return todo;
+    }
+    catch (error) {
+        console.log(error);
+        return error;
+    }
+}
+
+async function unshare(id, data) {
+    try {
+        const user = await usersController.getUserById(data.user);
+        if (!user) {
+            throw new Error('User not found');
+        }
+        if (!data.todo) {
+            throw new Error('Todo is required');
+        }
+        const todo = await todoController.gettodoById(id);
+        if (!todo) {
+            throw new Error('Todo not found');
+        }
+        // Ensure the user has access to update this todo
+        if (todo.user.toString() !== user._id.toString()) {
+            throw new Error('Unauthorized to update this todo');
+        }
+        // Update the todo
+        todo.sharedWith.pull(data.user);
+        await todo.save();
+        return todo;
+    }
+    catch (error) {
+        console.log(error);
+        return error;
+    }
+}
+
+async function getTodoByCategory(userId, category) {
+    console.log({ userId, category });
+    try {
+        const user = await usersController.getUserById(userId);
+        if (!user) {
+            throw new Error('User not found');
+        }
+        
+        const todos = await todoController.getAlltodo();
+        const filteredTodos = todos.filter(todo => todo.category === category && todo.user.toString() === userId);
+        console.log({ filteredTodos });
+        
+        return filteredTodos;
+    } catch (error) {
+        console.log(error);
+        return error;
+    }
+}
+
+
+// getTodoByCategory("General", "660115e6fc28d3e1ced32a6d")
+
+module.exports = { gettodoByUser, getOnetodo, create, update, deltodo, getCategories, share, unshare, getTodoByCategory };
+
+
+// async function create(data) {
+//     try {
+//         const user = await usersController.getUserById(data.user);
+//         if (!user) {
+//             throw error = new Error('no user')
+//         }
+//         if (!data.todo) {
+//             throw error = new Error('no todo')
+//         }
+//         const todo = await todoController.createtodo(data);
+//         user.todo.unshift(todo._id);
+//         await user.save();
+//         return todo;
+//     } catch (error) {
+//         console.log(error);
+//         return error;
+//     }
+// }
+
+
+// async function update(id, data) {
+//     try {
+//         const user = await usersController.getUserById(data.user);
+//         user.todo.filter(todo => todo.toString() !== id.toString());
+//         const todo = await todoController.updatetodo(id, data);
+//         user.todo.push(todo._id);
+//         await user.save();
+//         return todo;
+//     } catch (error) {
+//         console.log(error);
+//         return error;
+//     }
+// }
+
+// async function deltodo(id) {
+//     try {
+//         const todoToDelete = await todoController.gettodoById(id);
+//         const user = await usersController.getUserById(todoToDelete.user);
+
+//         console.log("user before deletion", user);
+
+//         // Use filter to create a new array without the todo to delete
+//         user.todo = user.todo.filter(todo => todo.toString() !== id.toString());
+
+//         console.log("user after deletion", user);
+
+//         const deletedtodo = await todoController.deletetodo(id);
+//         await user.save();
+
+//         return deletedtodo;
+//     } catch (error) {
+//         console.log(error);
+//         return error;
+//     }
+// }
+
+
